renaming config file
