+++ src/index.js
@@ -4,7 +4,7 @@ import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { Amplify } from "aws-amplify";
-import awsExports from "./components/authentication/auth-config";
+import awsExports from "./components/authentication/aws-export";
 Amplify.configure(awsExports);
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
